+++ packages/browser/src/ClientAuthentication.ts
@@ -33,10 +33,9 @@ import {
   IIssuerConfigFetcher,
   ISessionInternalInfo,
   ILoginOptions,
-  fetchJwks,
 } from "@inrupt/solid-client-authn-core";
 import { removeOidcQueryParam } from "@inrupt/oidc-client-ext";
-import { jwtVerify, parseJwk } from "@inrupt/jose-legacy-modules";
+import { jwtVerify, createRemoteJWKSet } from "@inrupt/jose-legacy-modules";
 import { EventEmitter } from "events";
 import { KEY_CURRENT_SESSION } from "./constant";
 
@@ -134,8 +133,8 @@ export default class ClientAuthentication {
     );
 
     try {
-      const jwk = await fetchJwks(issuerConfig.jwksUri, issuerConfig.issuer);
-      await jwtVerify(sessionInfo.idToken, await parseJwk(jwk), {
+      const jwks = createRemoteJWKSet(new URL(issuerConfig.jwksUri));
+      await jwtVerify(sessionInfo.idToken, jwks, {
         audience: sessionInfo.clientAppId,
         issuer: issuerConfig.issuer,
       });